refactor(logger): extract TestContext types and rename test context module
This refactoring improves code organization by separating type definitions
from implementation, making the codebase more maintainable and imports
more explicit.
